refactor: QueueController 리팩토링
