--- src/main/java/com/hhplus/concertreservation/queue/presentation/dto/response/CreateQueueResponse.java
@@ -1,13 +1,0 @@
-package com.hhplus.concertreservation.queue.presentation.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-
-import java.time.LocalDateTime;
-
-public record CreateQueueResponse (
-        String id,
-        String token,
-        @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-        LocalDateTime joinedAt
-) {
-}
